requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
scipy==1.12.0

# LLM
llama-cpp-python==0.2.19

# Speech processing
openai-whisper==20231117
sounddevice==0.4.6
soundfile==0.12.1

# Utilities
wget==3.2
gdown==4.7.3